requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
astroid==2.11.3
async-timeout==4.0.2
attrs==21.4.0
black==21.4b2
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.0.0
colorama==0.4.5
cryptography==36.0.2
cycler==0.11.0
datasets==2.0.0
dill==0.3.4
filelock==3.6.0
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.3.0
future==0.18.2
fvcore==0.1.5.post20220506
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.0
huggingface-hub==0.5.1
hydra-core==1.1.2
idna==3.3
importlib-metadata==4.11.3
install==1.3.5
iopath==0.1.9
isort==5.10.1
joblib==1.1.0
kiwisolver==1.4.2
lazy-object-proxy==1.7.1
Markdown==3.3.7
matplotlib==3.5.2
mccabe==0.7.0
multidict==6.0.2
multiprocess==0.70.12.2
mypy-extensions==0.4.3
numpy==1.22.3
oauthlib==3.2.0
omegaconf==2.1.2
packaging==21.3
pandas==1.4.2
pathspec==0.9.0
pdf2image==1.16.0
pdfminer.six==20220524
pdfplumber==0.7.4
Pillow==9.1.0
pip==21.2.3
platformdirs==2.5.2
portalocker==2.4.0
protobuf==3.20.1
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser==2.21
pydot==1.4.2
pylint==2.13.7
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.8
scikit-learn==1.1.2
scipy==1.8.0
sentencepiece==0.1.96
seqeval==1.2.2
setuptools==57.4.0
six==1.16.0
tabulate==0.8.9
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.1.0
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
torch==1.12.1
torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.9
Wand==0.6.7
Werkzeug==2.1.2
wheel==0.37.1
wrapt==1.14.0
xxhash==3.0.0
yacs==0.1.8
yarl==1.7.2
zipp==3.8.0
Jinja2==3.1.2
label-studio==0.0.0
MarkupSafe==2.1.1
nltk==3.7
processing-cv==0.0.0
processing-json==0.0.0
sentence-transformers==2.2.2
torch-cluster==1.6.0
torch-geometric==2.1.0.post1
torch-scatter==2.0.9
torch-sparse==0.6.15
torch-spline-conv==1.2.1
transformers==4.21.1
typed-ast==1.5.4
yapf==0.32.0
